Refactor table.py
